replace luahelper.h with LuaState

+++ src/render.h
@@ -123,9 +123,9 @@ struct RenderSettings
     uint32_t  mipmaps = 3;
     uint32_t  anisotropy = 0;
     bool antialias = false;
-    int8_t    antialias_samples = 0;
-    int8_t    texture_border = 8;
-    int8_t    z_depth = 16;
+    int       antialias_samples = 0;
+    int       texture_border = 8;
+    int       z_depth = 16;
     bool fog_enabled = true;
     GLfloat   fog_color[4]{0,0,0,1};
     float     fog_start_depth = 10000;
